[feat/#13] search post item 피드백 반영
